Add NewStringFromBytes function
The new function should be only used when the source of the string is a byte slice and
it is guaranteed that the bytes in the slice will not be modified or when the
immutability of the string stored inside this Variant is not required. In such cases
NewStringFromBytes(v) provides significant performance advantage over
NewString(string(v)) call which will create a copy of byte slice 'v'.